+++ utils/misc/app_state.h
